Organize chrome-scripts/assets into icons/ and css/ subfolders
