+++ src/chrome-scripts/features/content/readingList.ts
@@ -1,6 +1,6 @@
 import { createElement, createObserver, makeRequest } from '../../classes';
-import bookmarkMinusIcon from '../../assets/bookmark-minus.svg?url';
-import bookmarkPlusIcon from '../../assets/bookmark-plus.svg?url';
+import bookmarkMinusIcon from '../../assets/icons/bookmark-minus.svg?url';
+import bookmarkPlusIcon from '../../assets/icons/bookmark-plus.svg?url';
 
 export function readingList() {
   const observer = createObserver({
